+++ cpp/ycm/PythonSupport.h
@@ -32,7 +32,7 @@ YCM_EXPORT pybind11::list FilterAndSortCandidates(
   const pybind11::list &candidates,
   const std::string &candidate_property,
   const std::string &query,
-  const size_t max_candidates = 0 );
+  const long max_candidates = 0 );
 
 /// Given a Python object that's supposed to be "string-like", returns a UTF-8
 /// encoded std::string. Raises an exception if the object can't be converted to
